requirements.txt
flask==1.0.2
google-cloud-datastore==1.10.0